% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_set_data.R
\name{clean_and_set_data}
\alias{clean_and_set_data}
\title{Clean and Set all Data}
\usage{
clean_and_set_data(user_dir)
}
\value{
nothing. just sets the data and the sata is broken up into years. set_2017 up to set_2023
}
\description{
Clean and Set all Data
}
\examples{
clean_and_set_data(User/R/fullVRreport.xlsx)
}
