% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_trend_analysis_and_prediction.R
\name{county_trend_analysis_and_prediction}
\alias{county_trend_analysis_and_prediction}
\title{County Trends and Predictions}
\usage{
county_trend_analysis_and_prediction(county_name)
}
\arguments{
\item{county_name}{enter the FL county name you want to test in all caps in quotes}
}
\value{
all trends for each year and the prediction for the upcoming year based on the county.
}
\description{
County Trends and Predictions
}
\examples{
county_trend_analysis_and_prediction("ALACHUA")
}
