% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_year_pie.R
\name{county_year_pie}
\alias{county_year_pie}
\title{County Year Summary}
\usage{
county_year_pie(county_name, year)
}
\arguments{
\item{county_name}{name of the county in all caps and quotes}

\item{year}{year in format YYYY}
}
\value{
prints the pie chart for the year summary of the county.
}
\description{
County Year Summary
}
\examples{
county_year_pie("ALACHUA", 2019)
}
