% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_year_summary_pie.R
\name{total_year_summary_pie}
\alias{total_year_summary_pie}
\title{Total Year Pie Chart}
\usage{
total_year_summary_pie(year)
}
\arguments{
\item{year}{in format YYYY}
}
\value{
pie chart of division betweeen Democrats and Republicans throughout FL over a specific year.
}
\description{
Total Year Pie Chart
}
\examples{
total_year_summary_pie(2019)
}
